ssb: Clear RETRY_TIMEOUT in PCI Configuration for normal devices
The previous patch of this type fixed the case where a PCI device is behind
an SSB->PCI bridge as found on an embedded device. This one treats the normal
case of a "normal" Broadcom wireless PCI core on an SSB interconnect.

MMIO log traces obtained using the Broadcom wl hybrid driver show that
the RETRY_TIMEOUT register (0x41) in PCI configuration space is cleared
if non-zero. Similar code found in other drivers such as ipw2100 show
this operation is needed to keep PCI Tx retries from interfering with
C3 CPU state. There are no known cases where omission of this code has
caused a problem, but this patch is offered just in case such a situation
occurs.

Signed-off-by: Larry Finger <Larry.Finger@lwfinger.net>
Signed-off-by: John W. Linville <linville@tuxdriver.com>